feat(tracking): F1b float32 frame storage + team_id→category (schema, no weights; ADR-106)
Commit 1 of 2 (schema migration; the DGX retrain + version/CHANGELOG land in commit 2).

- float32 STORAGE for frame coords/kinematics; float64 COMPUTE (upcast at kernel
  boundaries) so drift = deterministic storage-rounding + numba/ADR-076 bit-identity
  intact. AST gate test_frame_coord_upcast_gate.
- Write-back-cast class: float64 results cast back to the column storage dtype at
  masked/scalar coord setitems (reflection/direction/utils/positioning/gkdv/
  restdefense/_model_eval) so pandas 3 does not raise LossySetitemError (ADR-057);
  forward-guarded by the pandas-3 CI leg, not a bespoke AST gate.
- team_id -> category (static/low-card); player_id stays Int64/object (dynamic
  post-build, ADR-103). id_compat._decat unwraps category value-neutrally; ADR-019
  category-axis gate (34 aggregators invariant); observed=True on team_id groupbys.
- Frames only -- SPADL actions stay float64; event-only models untouched.
- Part B (off-frame provenance) dropped; atomic-dedup residue nil in-repo.
- Gates: commit-1 model-load (F1B-SPEC-01), weights-dir classification (F1B-SPEC-02).

No weights, no version bump, no CHANGELOG (commit 2). No retrain from this commit.

+++ silly_kicks/tracking/_kernels.py
@@ -232,8 +232,8 @@ def _defenders_in_triangle_to_goal(
     by = np.full_like(ay, _GOAL_LEFT_POST_Y)
     cx = np.full_like(ax, _GOAL_X)
     cy = np.full_like(ay, _GOAL_RIGHT_POST_Y)
-    px = merged["x"].to_numpy()
-    py = merged["y"].to_numpy()
+    px = merged["x"].to_numpy(dtype="float64")
+    py = merged["y"].to_numpy(dtype="float64")
 
     def _sign(x1, y1, x2, y2, x3, y3):
         return (x1 - x3) * (y2 - y3) - (x2 - x3) * (y1 - y3)
@@ -441,8 +441,8 @@ def _pressure_andrienko(
 
     ax = merged["ax"].to_numpy()
     ay = merged["ay"].to_numpy()
-    px = merged["x"].to_numpy()
-    py = merged["y"].to_numpy()
+    px = merged["x"].to_numpy(dtype="float64")
+    py = merged["y"].to_numpy(dtype="float64")
 
     threat_dx = goal_x - ax
     threat_dy = goal_y - ay
@@ -520,8 +520,8 @@ def _pressure_link(
 
     ax = merged["ax"].to_numpy()
     ay = merged["ay"].to_numpy()
-    px = merged["x"].to_numpy()
-    py = merged["y"].to_numpy()
+    px = merged["x"].to_numpy(dtype="float64")
+    py = merged["y"].to_numpy(dtype="float64")
 
     presser_dx = px - ax
     presser_dy = py - ay
@@ -810,8 +810,8 @@ def _actor_pre_window_kernel(
     for aid, group in grouped:
         if len(group) < 2:
             continue
-        xs = group["x"].to_numpy()
-        ys = group["y"].to_numpy()
+        xs = group["x"].to_numpy(dtype="float64")
+        ys = group["y"].to_numpy(dtype="float64")
         dx = np.diff(xs)
         dy = np.diff(ys)
         arc = float(np.sqrt(dx * dx + dy * dy).sum())
